+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -53,16 +53,20 @@ public enum NoteTaskIdentifier
     [DisplayDescription("Confirm this transfer has authority to proceed")]
     ConfirmTransferHasAuthorityToProceed = 12,
 
+    [Description("confirm_date_academy_transferred")]
+    [DisplayDescription("Confirm the date the academy transferred")] // task heading 
+    ConfirmDateAcademyTransferred = 13,
+
     [Description("conditions_met")]
     [DisplayDescription("Confirm all conditions have been met")]
-    ConfirmAllConditionsMet = 13,
+    ConfirmAllConditionsMet = 14,
 
     [Description("receive_grant_payment_certificate")]
     [DisplayDescription("Receive declaration of expenditure certificate")]
-    ReceiveGrantPaymentCertificate = 14, 
+    ReceiveGrantPaymentCertificate = 15, 
 
     [Description("church_supplemental_agreement")]
     [DisplayDescription("Church supplemental agreement")]
-    ChurchSupplementalAgreement = 15,
+    ChurchSupplementalAgreement = 16,
     
 }